requirements.txt
kafka-python==2.2.7
tzdata==2025.2
streamlit==1.42.0
plotly==5.24.1
pandas==2.2.3
psycopg2-binary==2.9.9
numpy==2.1.3
cachetools>=5.3
requests>=2.32
pyspark==3.5.5
pytest>=8.0
testcontainers[kafka,postgresql]>=4.0
